Ratatui demo

+++ src/events.rs
@@ -3,7 +3,7 @@ pub enum Event {
     Power(bool),
     Volume(u8),
     Mute(bool),
-    Sleep(Option<u32>),
+    Sleep(Option<u8>),
     Input(String),
     Display(u8, String),
     Unknown(String),
